requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
xgboost==1.7.6
scikit-learn==1.3.2
pandas==2.2.0
numpy==1.26.3
sentence-transformers==2.3.1
joblib==1.3.2
shap==0.44.1
matplotlib==3.8.2
seaborn==0.13.1
python-multipart
python-dotenv
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4
pymysql==1.1.0
mysqlclient==2.2.4
aiomysql==0.2.0
streamlit==1.31.0
requests==2.31.0
pdfplumber==0.10.3
python-docx==1.1.0
spacy==3.7.2
nltk==3.8.1
faiss-cpu==1.7.4
aif360==0.6.1
torch==2.1.2
